+++ src/main/java/org/makechtec/financial/financialutils/cashflow/CashFlow.java
@@ -1,4 +1,7 @@
-package org.makechtec.financial.financialutils;
+package org.makechtec.financial.financialutils.cashflow;
+
+import org.makechtec.financial.financialutils.CompoundFactor;
+import org.makechtec.financial.financialutils.Rate;
 
 public class CashFlow{
