+++ src/test/java/org/makechtec/financial/financialutils/CashFlowTest.java
@@ -3,6 +3,7 @@ package org.makechtec.financial.financialutils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.makechtec.financial.financialutils.cashflow.CashFlow;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
