+++ src/test/java/org/makechtec/financial/financialutils/StreamOfCashFlowTest.java
@@ -2,6 +2,7 @@ package org.makechtec.financial.financialutils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.makechtec.financial.financialutils.cashflow.StreamOfCashFlow;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
